feat: Group member count
