+++ lib/leancloud.js
@@ -100,3 +100,7 @@ export const useObject = createResourceHook((className, objectId) => {
       .get({ abortSignal: abortController.signal }),
   };
 });
+
+export const useFunction = createResourceHook((functionName, params) => {
+  return cloud.run(functionName, params)
+})
